重构渲染模块组织架构
将渲染相关代码从分散的目录重新组织到统一的渲染模块中，提高代码结构的清晰度和可维护性。

  主要变更：
  - 删除旧的 `basic_drawing_elements/` 目录及其文件
  - 创建新的 `render/` 模块目录结构
  - 将基础图元移动到 `render/primitives/` 子目录：
    - `atlas_region.hpp`、`color_rgb.hpp`、`Vertex.hpp`
    - `glyph_atlas.hpp`（从原位置重命名移动）
  - 将渲染核心组件移动到 `render/` 目录：
    - `atlas_texture.hpp`（从basic_drawing_elements移动）
    - `text.hpp`、`triangle.hpp`（新增）
    - `vertex_batch_builder.hpp`（从frame/目录移动）
  - 更新对应的源文件位置和CMake配置

  重构目标：
  1. 集中管理所有渲染相关代码
  2. 提供更清晰的模块分层（primitives基础图元 → render核心渲染）
  3. 为未来渲染功能扩展提供更好的基础结构

  这个提交信息清晰描述了：
  1. 做了什么：渲染模块重构
  2. 具体变更：文件移动、删除和新增的详细说明
  3. 重构目的：提高代码组织性和可维护性
  4. 未来意义：为扩展奠定基础

+++ code/src/core/frame/window.cpp
@@ -1,9 +1,8 @@
 #include <core/frame/window.hpp>
 #include <core/frame/window_controller.hpp>
-#include <core/basic_drawing_elements/atlas_region.hpp>
+#include <core/render/primitives/atlas_region.hpp>
 #include <SDL3_image/SDL_image.h>
 #include <utility>
-#include <chrono>
 #include <ranges>
 
 namespace dao {
@@ -135,7 +134,7 @@ namespace dao {
             m_windowFlags);
         m_renderer = SDL_CreateRenderer(m_window, "direct3d11");
         SDL_SetRenderVSync(m_renderer, 0);
-        m_id = SDL_GetWindowID(m_window);
+        m_id = static_cast<i32>(SDL_GetWindowID(m_window));
     }
 
 
@@ -156,7 +155,7 @@ namespace dao {
     }
 
     void Window::setPosition(const i32 x, const i32 y) const {
-        SDL_SetWindowPosition(m_window, x,y);
+        SDL_SetWindowPosition(m_window, x, y);
     }
 
     void Window::movePosition(const i32 x, const i32 y) const {
